requirements.txt
affine==2.0.0.post1
click==6.7
click-plugins==1.0.3
cligj==0.4.0
enum34==1.1.6
geojson==1.0.7
numpy==1.12.0
pandas==0.19.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-dotenv==0.6.3
pytz==2016.10
rasterio==0.36.0
requests==2.4.3
scipy==0.18.1
Shapely==1.5.17.post1
six==1.10.0
snuggs==1.4.1
